requirement.txt
flask
torch
diffusers
rembg
pillow
